learned to interact

+++ hardhat-simple-storage/contracts/SimpleStorage.sol
@@ -10,7 +10,7 @@ contract SimpleStorage {
     //storage state variables. permenant variables that can be modified.
     int public data;
     string public sName;
-    uint public age;
+    uint256 public age;
     // Play ex;
     address public deployedContractAddress;
 
@@ -22,7 +22,7 @@ contract SimpleStorage {
     //Struct
     struct People {
         string name;
-        uint age;
+        uint256 age;
         bool married;
     }
 
@@ -30,13 +30,13 @@ contract SimpleStorage {
     People[] public peopleInfo;
 
     //Mappping (dictionary)
-    mapping(string => uint) public peopleData;
+    mapping(string => uint256) public peopleData;
 
     //callData, memory -> variable exist temporary & Storage -> permenant storage (state variables)
 
     function addData(
         string memory _peopleName,
-        uint _age,
+        uint256 _age,
         bool _isMarried
     ) public {
         //alternative of the above.
@@ -52,12 +52,11 @@ contract SimpleStorage {
         peopleData[_peopleName] = _age;
     }
 
-    function getPeopleInfoLength() public view returns (uint) {
+    function getPeopleInfoLength() public view returns (uint256) {
         return peopleInfo.length;
-
     }
 
-    function setData(string memory _sName, uint _age) public virtual {
+    function setData(string memory _sName, uint256 _age) public virtual {
         sName = _sName;
         age = _age;
     }
@@ -75,11 +74,13 @@ contract SimpleStorage {
         return deployedContractAddress;
     }
 
-    function getInfo(string memory _name) external pure returns(string memory){
+    function getInfo(
+        string memory _name
+    ) external pure returns (string memory) {
         return _name;
     }
 
-    function getDetails(string memory _name) public view returns(uint){
+    function getDetails(string memory _name) public view returns (uint256) {
         return peopleData[_name];
     }
 
@@ -87,7 +88,7 @@ contract SimpleStorage {
 
     function displayAllData(
         string memory _userData
-    ) public  view returns (uint, People memory, string memory, uint) {
+    ) public view returns (uint256, People memory, string memory, uint256) {
         return (peopleData[_userData], People("Sumit", 30, true), "no data", 1);
     }
 }
